requirements.txt
streamlit
openai==0.28
python-docx
